+++ spec/alpha_card/transactions/auth_spec.rb
@@ -17,7 +17,7 @@ describe AlphaCard::Auth do
   context 'with valid attributes' do
     it 'has valid request params' do
       expected_params = {
-        ccexp: '09/2017',
+        ccexp: card_exp,
         ccnumber: '4111111111111111',
         amount: '5.00',
         payment: 'creditcard',