+++ spec/alpha_card/transactions/validate_spec.rb
@@ -1,54 +1,51 @@
 require 'spec_helper'
 
-describe AlphaCard::Credit do
+describe AlphaCard::Validate do
   let(:order) { AlphaCard::Order.new(id: '1', description: 'Test') }
   let(:card_exp) { "#{'%02d' % Time.now.month}/#{Time.now.year.next}" }
 
   context 'with valid attributes' do
-    let(:credit) { AlphaCard::Credit.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '5.00') }
+    let(:validate) { AlphaCard::Validate.new(card_expiration_date: card_exp, card_number: '4111111111111111') }
 
     it 'has valid request params' do
       expected_params = {
         ccexp: card_exp,
         ccnumber: '4111111111111111',
-        amount: '5.00',
         payment: 'creditcard',
         orderid: '1',
         orderdescription: 'Test',
-        type: 'credit',
+        type: 'validate',
       }
 
-      expect(credit.send(:params_for_sale, order)).to eq(expected_params)
+      expect(validate.send(:params_for_sale, order)).to eq(expected_params)
     end
 
-    it 'successfully creates the credit' do
-      resp  = credit.create(order)
-      puts resp.inspect
-      expect(resp.success?).to be_truthy
+    it 'successfully processed' do
+      response = validate.create(order)
+      expect(response.success?).to be_truthy
+    end
+  end
+
+  context 'with amount' do
+    let(:validate) { described_class.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '1.22') }
+
+    it 'must ignore it' do
+      expect(validate.send(:params_for_sale, order)).not_to include(:amount)
     end
   end
 
-  context 'with invalid amount' do
-    let(:credit) { AlphaCard::Credit.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '0.00') }
-    let(:response) { credit.process(order) }
-
-    it 'returns an error' do
-      expect(response.error?).to be_truthy
-      expect(response.text).to include('Invalid amount')
-    end
-  end
 
   context 'without attributes' do
-    let(:credit) { AlphaCard::Credit.new }
+    let(:validate) { described_class.new }
 
     it 'raises an InvalidObjectError exception' do
-      expect { credit.create(order) }.to raise_error(AlphaCard::InvalidObjectError)
+      expect { validate.create(order) }.to raise_error(AlphaCard::InvalidObjectError)
     end
   end
 
   context 'with invalid account credentials' do
-    let(:credit) { AlphaCard::Credit.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '5.00') }
-    let(:response) { credit.process(order, username: 'demo', password: 'Invalid password') }
+    let(:validate) { described_class.new(card_expiration_date: card_exp, card_number: '4111111111111111') }
+    let(:response) { validate.process(order, username: 'demo', password: 'Invalid password') }
 
     it 'returns an error' do
       expect(response.error?).to be_truthy
@@ -57,10 +54,10 @@ describe AlphaCard::Credit do
   end
 
   context 'with blank account credentials' do
-    let(:credit) { AlphaCard::Credit.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '5.00') }
+    let(:validate) { described_class.new(card_expiration_date: card_exp, card_number: '4111111111111111') }
 
     it 'raises an ArgumentError' do
-      expect { credit.process(order, username: nil, password: '') }.to raise_error(ArgumentError) do |e|
+      expect { validate.process(order, username: nil, password: '') }.to raise_error(ArgumentError) do |e|
         expect(e.message).to include('You must pass a Hash with Account credentials!')
       end
     end
@@ -90,11 +87,11 @@ describe AlphaCard::Credit do
     end
 
     context 'with request exception' do
-      let(:credit) { AlphaCard::Credit.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '5.00') }
+      let(:validate) { described_class.new(card_expiration_date: card_exp, card_number: '4111111111111111') }
 
       it 'handles an error' do
         AlphaCard.api_base = 'https://not-existing.com'
-        expect { credit.create(order) }.to raise_error(AlphaCard::APIConnectionError)
+        expect { validate.create(order) }.to raise_error(AlphaCard::APIConnectionError)
 
         AlphaCard.api_base = 'https://secure.alphacardgateway.com/api/transact.php'
       end